% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_cor.R
\name{net_cor}
\alias{net_cor}
\title{Estimate network correlation and social differentiation for interaction
networks.}
\usage{
net_cor(X, D, directed = FALSE, ci = 0.95, show_diagnostic = TRUE)
}
\arguments{
\item{X}{n x n integer-valued matrix with zero-valued diagonals where each
entry (i,j) gives the number of observed interactions between i and j.}

\item{D}{n x n real-valued positive matrix with zero-valued diagonals where
each entry (i, j) gives the amount of time spent sampling i and j.}

\item{directed}{TRUE/FALSE indicating if the network is directed or not.}

\item{ci}{Real value between 0 and 1 giving the width of the confidence
interval.}

\item{show_diagnostic}{TRUE/FALSE indicating if to show the diagnostic QQ
plot.}
}
\value{
Summary table of network properties including network correlation and
  social differentiation.
}
\description{
Estimates properties of an estimated network with `n` nodes including network
correlation and social differentiation.
}
\details{
By default this function plots a QQ diagnostic plot to indicate how well the
raw data fit the Gamma-Poisson model assumed by this method. If most of the
points are close to the diagonal line, a good fit can be assumed. Sometimes
the tails of the distribution may curve away from the line, but provided this
is limited to the tails it is unlikely to significantly affect the results.

Note: This function will only work with interaction rate networks.
}
\examples{
# Simulate observation data.
sim_data <- simulate_data_gp(20, 10, 0.25, 0.5)
X <- sim_data$X
D <- sim_data$D

# Estimate network sampling properties.
net_cor(X, D)
}
