% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{simulate_data_gp}
\alias{simulate_data_gp}
\title{Simulate networks based on Gamma-Poisson model.}
\usage{
simulate_data_gp(n, mean_sampling, S, mu, directed = FALSE)
}
\arguments{
\item{n}{Integer greater than zero describing number of nodes in the network.}

\item{mean_sampling}{Real number greater than zero describing the mean amount
of time spent sampling each dyad.}

\item{S}{Real number greater than zero describing the desired social
differentiation of the network.}

\item{mu}{Real number greater than zero describing the mean number of
interactions observed per dyad.}

\item{directed}{TRUE/FALSE indicating if the network should be directed.}
}
\value{
A list of three `n x n` matrices. `X` is the interaction observation
  matrix, `D` is the sampling time, and `A` is a matrix of the true
  interaction rates.
}
\description{
Simulates networks based on the Gamma-Poisson model and outputs corresponding
adjacency matrices.
}
